Improve Adapters (parse_value) and fix bugs & typos
