+++ dspy/adapters/utils.py
@@ -1,9 +1,11 @@
+import ast
+import enum
 import json
-from typing import Any, List, Literal, Union, get_args, get_origin
+import json_repair
 
 from pydantic import TypeAdapter
 from pydantic.fields import FieldInfo
-
+from typing import Any, List, Literal, Union, get_args, get_origin
 
 
 def serialize_for_json(value: Any) -> Any:
@@ -82,6 +84,27 @@ def find_enum_member(enum, identifier):
         return enum[identifier]
 
     raise ValueError(f"{identifier} is not a valid name or value for the enum {enum.__name__}")
+
+
+
+def parse_value(value, annotation):
+    if annotation is str:
+        return str(value)
+
+    if isinstance(annotation, enum.EnumMeta):
+        return find_enum_member(annotation, value)
+
+    if not isinstance(value, str):
+        return TypeAdapter(annotation).validate_python(value)
+
+    candidate = json_repair.loads(value) # json_repair.loads returns "" on failure.
+    if candidate == "" and value != "":
+        try:
+            candidate = ast.literal_eval(value)
+        except (ValueError, SyntaxError):
+            candidate = value
+
+    return TypeAdapter(annotation).validate_python(candidate)
 
 
 def get_annotation_name(annotation):
