+++ dspy/evaluate/metrics.py
@@ -3,14 +3,20 @@
 
 def _passage_match(passages: list[str], answers: list[str]) -> bool:
     """Returns True if any of the passages contains the answer."""
+    from dspy.dsp.utils import DPR_normalize, has_answer, normalize_text
 
-    from dspy.dsp.utils import passage_has_answers
+    def passage_has_answers(passage: str, answers: list[str]) -> bool:
+        """Returns True if the passage contains the answer."""
+        return has_answer(
+            tokenized_answers=[DPR_normalize(normalize_text(ans)) for ans in answers],
+            text=normalize_text(passage),
+        )
+
     return any(passage_has_answers(psg, answers) for psg in passages)
 
 
 def _answer_match(prediction, answers, frac=1.0):
     """Returns True if the prediction matches any of the answers."""
-
     from dspy.dsp.utils import EM, F1
 
     if frac >= 1.0:
